+++ module/web/src/router/standard.ts
@@ -1,13 +1,16 @@
 import router from 'find-my-way';
 
 import { AppError, castTo } from '@travetto/runtime';
-import { Injectable } from '@travetto/di';
+import { Inject, Injectable } from '@travetto/di';
 
 import { EndpointConfig } from '../registry/types.ts';
 
 import { WebResponse } from '../types/response.ts';
 import { HTTP_METHODS, HttpMethod } from '../types/core.ts';
 import { WebFilterContext } from '../types.ts';
+import { WebConfig } from '../config/web.ts';
+import { WebCommonUtil } from '../util/common.ts';
+
 import { BaseWebRouter } from './base.ts';
 
 /**
@@ -15,6 +18,9 @@ import { BaseWebRouter } from './base.ts';
  */
 @Injectable()
 export class StandardWebRouter extends BaseWebRouter {
+
+  @Inject()
+  config: WebConfig;
 
   #cache = new Map<Function, EndpointConfig>();
   raw = router();
@@ -45,6 +51,6 @@ export class StandardWebRouter extends BaseWebRouter {
       throw new AppError(`Unknown route ${req.method} ${req.path}`, { category: 'notfound' });
     }
     Object.assign(req, { params });
-    return endpoint.filter!({ req });
+    return endpoint.filter!({ req: WebCommonUtil.secureRequest(req, this.config.trustProxy) });
   }
 }
